refactor: remove deprecated address define
