+++ packages/permit2-sdk/src/constants.ts
@@ -1,28 +1,38 @@
 import { ChainId } from '@pancakeswap/chains'
 import { Address } from 'viem'
 
-const PERMIT2_ADDRESSES: { [key: number]: Address } = {
-  [ChainId.ETHEREUM]: '0x0000000000000000000000000000000000000000',
+const PERMIT2_ADDRESSES: Record<ChainId, Address> = {
+  [ChainId.ETHEREUM]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
   [ChainId.GOERLI]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+
   [ChainId.BSC]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
   [ChainId.BSC_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.SCROLL_SEPOLIA]: '0x0000000000000000000000000000000000000000',
-  [ChainId.ARBITRUM_ONE]: '0x0000000000000000000000000000000000000000',
-  [ChainId.ARBITRUM_GOERLI]: '0xCaCC5DF33e3AF664158eaaB87bc8f282C674ec25',
-  [ChainId.BASE_TESTNET]: '0xCaCC5DF33e3AF664158eaaB87bc8f282C674ec25',
-  [ChainId.BASE]: '0x0000000000000000000000000000000000000000',
-  [ChainId.POLYGON_ZKEVM]: '0x0000000000000000000000000000000000000000',
-  [ChainId.POLYGON_ZKEVM_TESTNET]: '0xCaCC5DF33e3AF664158eaaB87bc8f282C674ec25',
-  [ChainId.LINEA]: '0x0000000000000000000000000000000000000000',
-  [ChainId.LINEA_TESTNET]: '0xCeccEEA1ee8f7420aB5A0daFA56faF39Ee794933',
-  [ChainId.OPBNB_TESTNET]: '0xCaCC5DF33e3AF664158eaaB87bc8f282C674ec25',
-  [ChainId.ZKSYNC]: '0x0000000000000000000000000000000000000000',
-  [ChainId.ZKSYNC_TESTNET]: '0x0000000000000000000000000000000000000000',
+
+  // [ChainId.SCROLL]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+  [ChainId.SCROLL_SEPOLIA]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+
+  [ChainId.ARBITRUM_ONE]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+  [ChainId.ARBITRUM_GOERLI]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+
+  [ChainId.BASE]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+  [ChainId.BASE_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+
+  [ChainId.POLYGON_ZKEVM]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+  [ChainId.POLYGON_ZKEVM_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+
+  [ChainId.LINEA]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+  [ChainId.LINEA_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+
+  [ChainId.ZKSYNC]: '0x686FD50007EaA636F01154d660b96110B6bFe351',
+  [ChainId.ZKSYNC_TESTNET]: '0xaf321b731E65715DdbFDa73A066E00BB28345709',
+
+  [ChainId.OPBNB]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
+  [ChainId.OPBNB_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
 }
 
 export const getPermit2Address = (chainId: ChainId | undefined): Address => {
   if (chainId === undefined) return PERMIT2_ADDRESSES[ChainId.BSC]
-  if (!(chainId in PERMIT2_ADDRESSES)) throw new Error(`Universal Router not deployed on chain ${chainId}`)
+  if (!(chainId in PERMIT2_ADDRESSES)) throw new Error(`Permit2 Contract not deployed on chain ${chainId}`)
   return PERMIT2_ADDRESSES[chainId]
 }
