+++ packages/universal-router-sdk/src/index.ts
@@ -1,4 +1,4 @@
 export { PancakeSwapUniversalRouter } from './SwapRouterPancake'
 export * from './entities'
 export * from './entities/types'
-export { getUniversalRouterAddress, getPermit2Address, ROUTER_AS_RECIPIENT } from './utils/constants'
+export { getUniversalRouterAddress, ROUTER_AS_RECIPIENT } from './utils/constants'
