+++ packages/universal-router-sdk/src/utils/constants.ts
@@ -30,43 +30,9 @@ const UNIVERSAL_ROUTER_ADDRESSES: Record<ChainId, Address> = {
   [ChainId.OPBNB_TESTNET]: '0xa8EEA7aa6620712524d18D742821848e55E773B5',
 }
 
-const PERMIT2_ADDRESSES: Record<ChainId, Address> = {
-  [ChainId.ETHEREUM]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.GOERLI]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  [ChainId.BSC]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.BSC_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  // [ChainId.SCROLL]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.SCROLL_SEPOLIA]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  [ChainId.ARBITRUM_ONE]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.ARBITRUM_GOERLI]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  [ChainId.BASE]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.BASE_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  [ChainId.POLYGON_ZKEVM]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.POLYGON_ZKEVM_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  [ChainId.LINEA]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.LINEA_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-
-  [ChainId.ZKSYNC]: '0x686FD50007EaA636F01154d660b96110B6bFe351',
-  [ChainId.ZKSYNC_TESTNET]: '0xaf321b731E65715DdbFDa73A066E00BB28345709',
-
-  [ChainId.OPBNB]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-  [ChainId.OPBNB_TESTNET]: '0x31c2F6fcFf4F8759b3Bd5Bf0e1084A055615c768',
-}
-
 export const getUniversalRouterAddress = (chainId: ChainId): Address => {
   if (!(chainId in UNIVERSAL_ROUTER_ADDRESSES)) throw new Error(`Universal Router not deployed on chain ${chainId}`)
   return UNIVERSAL_ROUTER_ADDRESSES[chainId]
-}
-
-export const getPermit2Address = (chainId: ChainId): Address => {
-  if (!(chainId in PERMIT2_ADDRESSES)) throw new Error(`Permit2 Address not deployed on chain ${chainId}`)
-  return PERMIT2_ADDRESSES[chainId]
 }
 
 export const CONTRACT_BALANCE = 2n ** 255n
